% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_aggregation.R
\name{aggregate_pe}
\alias{aggregate_pe}
\title{Summarize peptide quantity to protein quantity}
\usage{
aggregate_pe(
  pe,
  aggrefun = c("RobustSummary", "medianPolish", "totalMean", "directLFQ"),
  aggregate_Peptide_Type = c("Unique + Razor", "Unique"),
  fcol,
  i = "peptideNorm",
  reserve = "Gene.names"
)
}
\arguments{
\item{pe}{A QFeatures object,
contains the normalized peptide assay}

\item{aggrefun}{A function used for quantitative feature aggregation.
It can be a character in "RobustSummary","directLFQ","medianPolish","totalMean" or
a function. Method directLFQ is required package Rdirectlfq.
Details see \code{\link[QFeatures]{aggregateFeatures}}}

\item{aggregate_Peptide_Type}{Character in "Unique + Razor" or "Unique".
Use what kind of peptides to summarise proteins. If choose "Unique", return output just save unique peptides in smallest proteingroups.}

\item{fcol}{Character(1), defining how to summarise the features. Exist in \code{rowData(pe)}.}

\item{i}{Character(1), name of the assay to be aggregated.}

\item{reserve}{Character, the column(s) which will reserve after aggregate, such as the columns store protein information can.}
}
\value{
A QFeatures object with a new protein aggregation assay.
}
\description{
Aggregation peptide to protein vis 'aggrefun'.
}
\examples{
\dontrun{
# Load example peptide data
data(Silicosis_peptide)
ecols <- grep("Intensity.", colnames(Silicosis_peptide), value = TRUE)

# construct QFeatures object
pe_peptides <- make_pe_parse(Silicosis_peptide, columns = ecols, remove_prefix = TRUE, log2transform = TRUE,mode = "delim")

# Filter, imputation
pe <- filter_pe(pe_peptides, thr = 1,fraction = 0.4, filter_formula = ~ Reverse != '+' & Potential.contaminant !="+" )
pe <- impute_pe(pe, fun = "QRILC", name = "peptideImp")

# Normalization
pe <- normalize_pe(pe,method = "quantiles", i = "peptideImp", name = "peptideNorm")

# Summarize peptide value to protein quantity
protein_pe <- DEP2::aggregate_pe(pe, fcol = "Proteins", i = "peptideNorm")
class(protein_pe)

}
}
