% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_enrichment.R
\name{annoSpecies_df}
\alias{annoSpecies_df}
\title{Species information table}
\usage{
annoSpecies_df()
}
\value{
a data.frame
}
\description{
Record the species names IDs and their org database package.
}
