% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Depends.R
\name{check_organismDB_depends}
\alias{check_organismDB_depends}
\title{Check organism annotation package}
\usage{
check_organismDB_depends(organism, install = T)
}
\arguments{
\item{organism}{Character(1), the organism name, could be one of c(
"Chicken", "Human", "Mouse", "Fly", "Rat", "Yeast"). Other species with
\href{http://bioconductor.org/packages/release/BiocViews.html#___OrgDb}{OrgDb} are also optional.}

\item{install}{logic(1). Whether install if organism annotation db package is missed.}
}
\description{
Check and install organism annotation package(OrgDb from Bioconductor).
}
