% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Depends.R
\name{check_PPI_depends}
\alias{check_PPI_depends}
\title{Check required packages for PPI analysis.}
\usage{
check_PPI_depends(install = T)
}
\arguments{
\item{install}{logic(1). Whether install the missing packages.}
}
\description{
Check the depend packages for PPI analysis.
}
