% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Depends.R
\name{check_RNAseq_depends}
\alias{check_RNAseq_depends}
\title{Check required packages for DESeq2.}
\usage{
check_RNAseq_depends(install = T)
}
\arguments{
\item{install}{logic(1). Whether install the missing packages.}
}
\description{
Check the depend packages for RNA-Seq analysis workflow based on DESeq2
}
