% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_protein.R
\name{clean_character}
\alias{clean_character}
\title{Clean characters in expression columns}
\usage{
clean_character(express_assay)
}
\arguments{
\item{express_assay}{data.frame or matrix, the expression assay.}
}
\value{
a numeric matrix
}
\description{
If the class of expression contain characters(e.g. 'NA.', 'filtered'), run Clean characters to replace character
by NA, and transfor expression value to numeric
}
\examples{
expression_assay = data.frame(A = c("1165799.5","51419028","1.9e6","2213111.25"),
                              B = c("1455111","Filtered","2021020.375","NA"),
                              c = c(261344, 49840292, 2388750, 1753475.625))
str(expression_assay)
(clean_character(expression_assay))

}
