% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_peptide.R
\name{correct_PTM_by_Protein}
\alias{correct_PTM_by_Protein}
\title{correct_PTM_by_Protein}
\usage{
correct_PTM_by_Protein(
  enriched_peptide,
  relative_protein,
  correct_key = c("name", "ID"),
  correct_level = c("condition", "replicate"),
  unidentified_treatment = c("retain", "remove")
)
}
\arguments{
\item{enriched_peptide}{The SummarizedExperiment object of modified-relative proteomics}

\item{relative_protein}{The SummarizedExperiment object of the relative global proteomics}

\item{correct_key}{The key identifier column to match PTM peptides and proteins, one of "name", "ID"}

\item{correct_level}{Character(1), how to match experiment design, should be one of "condition" or "replicate"}

\item{unidentified_treatment}{Character(1), one of "retain", "remove", how to treat the PTM peptides whose proteins are
unidentified in the relative global proteomics}
}
\value{
A modified-relative proteomics SummarizedExperiment object with adjusted quantities
}
\description{
Adjust modified-enrich peptide quantities (PTM relative proteomics) by a relative global proteomics data via matching
\code{correct_key} and \code{correct_level}
}
