% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_rna.R
\docType{class}
\name{DEGdata-class}
\alias{DEGdata-class}
\title{Class "DEGdata"
This class store the test result from DESeq2}
\description{
Class "DEGdata"
This class store the test result from DESeq2
}
\section{Slots}{

\describe{
\item{\code{geneinfo}}{data.frame. Store gene IDs and SYMBOL information.}

\item{\code{ntf}}{matrix. Log2 transform assay}

\item{\code{rlg}}{matrix. rlog transform assay}

\item{\code{test_result}}{DataFrame. Test result from DESeq2}
}}

