% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP2.R
\docType{package}
\name{DEP2}
\alias{DEP2-package}
\alias{DEP2}
\title{DEP2: An analysis toolkits for mass spectrometry based proteomics data.}
\description{
DEP2 provides an comprehensive analysis workflow for mass spectrometry based proteomics data, developed from the previous package DEP. This package provided differential expression/enrichment analysis pipelines for various data including protein-level quantity(e.g. proteingroup), peptide-level quantity data and modification-specific proteomics (quantities of modified peptides). I nherited from DEP, DEP2 provides functions for data processing, hypothesis testing (via limma) and visualization. The peptide to protein aggregation strategy is integrated into workflows. To reduced the barrier in omics analysis, downstream functional exploration are packaged as suites including functional enrichment, timecourse cluster and protein-protein interaction network. DEP2 also contains an easy-used shiny application designed under modularization.
}
\author{
\strong{Maintainer}: Zhenhuan Feng \email{feng_zhenhuan@gibh.ac.cn}

Authors:
\itemize{
  \item Xiaofei Zhang \email{zhang_xiaofei@gibh.ac.cn}
}

}
