% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_protein.R, R/DEP_rna.R
\name{add_rejections}
\alias{add_rejections}
\alias{add_rejections,SummarizedExperiment-method}
\alias{add_rejections,DEGdata-method}
\title{Mark significant proteins}
\usage{
\S4method{add_rejections}{SummarizedExperiment}(diff, alpha, lfc, thresholdmethod, curvature, x0_fold)

\S4method{add_rejections}{DEGdata}(diff, alpha, lfc, thresholdmethod, curvature, x0_fold)
}
\arguments{
\item{alpha}{Numeric(1),
Sets the threshold for the adjusted P value.}

\item{lfc}{Numeric(1),
Sets the threshold for the log2 fold change.}

\item{thresholdmethod}{Character(1),
The thresholdmethod to decide significant.should be one of "intersect" ,"curve". If thresholdmethod = "curve", cutoff lines is the curves with function y = c/(x-x0). c is the curvature, and x0 is the minimum fold change}

\item{curvature}{Numeric(1),
Sets the curvature for the curve cutoff lines}

\item{x0_fold}{Numeric(1),
decide the x0 ratio to the standard deviations of L2FC. The x0 usually is set to 1(medium confidence) or 2(high confidence) standard deviations.}
}
\value{
A SummarizedExperiment object
annotated with logical columns indicating significant proteins.
}
\description{
\code{add_rejections} marks significant proteins based on defined cutoffs.
}
\examples{
# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter, normalize and impute missing values
filt <- filter_missval(se, thr = 0)
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.01)

# Test for differentially expressed proteins
diff <- test_diff(imputed, "control", "Ctrl")
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)
}
