% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SE.R
\name{filter_se}
\alias{filter_se}
\title{Filter SummarizedExperiment on missing values or formula}
\usage{
filter_se(
  se,
  thr = NULL,
  missnum = NULL,
  fraction = NULL,
  filter_formula = NULL,
  rowsum_threshold = NULL,
  return_keeprows = FALSE
)
}
\arguments{
\item{se}{SummarizedExperiment,
Proteomics data (output from \code{\link{make_se}()} or
\code{\link{make_se_parse}()}).}

\item{thr}{Integer(1),
Sets the threshold for the allowed max number of missing values
in at least one condition. At least in one condition, missing values number < or = 'thr'.}

\item{missnum}{Integer(1),
Sets the threshold for the allowed max number among all samples,
missing values number < or = 'missnum' in each row.}

\item{fraction}{A numeric from 0 to 1, threshold of missing occupancy of each row}

\item{filter_formula}{Formula or character(1).
A filter expression, written as a formula.}

\item{rowsum_threshold}{A numeric, row sum of intensity(abundance) must larger than it.}

\item{return_keeprows}{Logical(1), default FALSE, return the trimmed object.
If TURE return the rows number that pass through filter.}
}
\value{
A filtered SummarizedExperiment object,
or a vector of keep rows if \code{return_keeprows} is TRUE
}
\description{
\code{filter_se} filters a SummarizedExperiment object based on missing values in assay
and provides formula.
The dataset is filtered for proteins that have a maximum of
'thr' missing values in at least one condition; or a maximum of 'missnum' in all samples;
or a filter_formula
}
\examples{
# Load example
data <- UbiLength
data <- data[data$Reverse != "+" & data$Potential.contaminant != "+",]
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))
exp_design <- UbiLength_ExpDesign
se <- make_se(data_unique, columns, exp_design)

# Filter
stringent_filter <- filter_se(se, thr = 0, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
less_stringent_filter <- filter_se(se, thr = 1, missnum = 2, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
keep_rows <- filter_se(se, thr = 0, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+", return_keeprows = TRUE)
}
