% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_result.R
\name{get_contrast}
\alias{get_contrast}
\title{Get contrast(s) from SummarizedExperiment or DEGdata}
\usage{
get_contrast(object)
}
\arguments{
\item{object}{A SummarizedExperiment or DEGdata object output from \link{test_diff}()}
}
\value{
Character, the existing contrast(s) in object, which is introduced in the test_diff step.
}
\description{
Get contrast(s) from SummarizedExperiment or DEGdata
}
