% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DEP1.R
\name{get_df_wide}
\alias{get_df_wide}
\title{Generate a wide data.frame from a SummarizedExperiment}
\usage{
get_df_wide(se)
}
\arguments{
\item{se}{SummarizedExperiment or DEGdata object.}
}
\value{
A data.frame object
containing all data in a wide format(combine the assay and rowData),
where each row represents a protein/gene.
}
\description{
\code{get_df_wide} generate a wide data.frame from a SummarizedExperiment.
}
\examples{
# Load example
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols, mode = "delim", sep = "_")
# Filter and normalize
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinDet")
diff <- DEP2::test_diff(imputed,type = "control", control = "PBS")
dep <- add_rejections(diff)
get_df_wide(dep)

}
