% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_enrichment.R
\name{get_ORA_result}
\alias{get_ORA_result}
\title{Extract significant enrichment terms base on giving threshold}
\usage{
get_ORA_result(
  ORA_enrichment,
  ont = NULL,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  simplify = FALSE,
  simplify.cutoff = 0.7,
  simplify.measure = c("Wang", "Resnik", "Lin", "Rel", "Jiang"),
  simplify.semData = NULL,
  return_table = F
)
}
\arguments{
\item{ont}{One of "ALL", "BP", "MF", "CC"}

\item{pvalueCutoff}{Numeric(1), the p.value cutoff on enrichment result}

\item{qvalueCutoff}{Numeric(1), the qvalue cutoff on enrichment tests}

\item{simplify}{Logical(1), if simplify GO terms by \code{\link[clusterProfiler]{simplify}}}

\item{simplify.cutoff}{Numeric(1), the cutoff value transmitted to  \code{\link[clusterProfiler]{simplify}}}

\item{return_table}{Logical(1), if true return a enrichResult or a result table}

\item{reat}{The output from test_ORA}
}
\value{
A enrichResult/compareClusterResult of significant enrichment,
or a result table of the significant enrichment if return_table is TRUE.
}
\description{
\code{get_ORA_result} filter the enrichment result
from \code{test_ORA} though certain threshold.
}
\examples{
\dontrun{
# Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Differential test
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols,mode = "delim")
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.05)
diff <- test_diff(imputed, type = "control", control  = c("PBS"), fdr.type = "Storey's qvalue")

res_ora <- test_ORA(dep, contrasts = "W4_vs_PBS", species = "Mouse",type = "GO")
res_ora2 <- get_ORA_result(res_ora)
}
}
