% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DEP1.R
\name{get_prefix}
\alias{get_prefix}
\title{Obtain the longest common prefix}
\usage{
get_prefix(words)
}
\arguments{
\item{words}{Character vector,
A list of words.}
}
\value{
A character vector containing the prefix.
}
\description{
\code{get_prefix} returns the longest common prefix
of the supplied words.
}
\examples{
# Load example
data(Silicosis_pg)
data <- Silicosis_pg
names <- grep("LFQ.", colnames(data), value = TRUE)

# Get prefix
get_prefix(names)
}
