% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_result.R
\name{get_results}
\alias{get_results}
\title{Generate a results table}
\usage{
get_results(object)
}
\arguments{
\item{object}{SummarizedExperiment or DEGdata,
(output from \code{\link{test_diff}()} and \code{\link{add_rejections}()}).}
}
\value{
A data.frame object
containing all results variables from the performed analysis.
}
\description{
\code{get_results} generates a results table from a proteomics or RNA-seq dataset
on which differential analysis was performed.
}
\examples{
#Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Differential test
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols,mode = "delim")
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.05)
diff <- test_diff(imputed, type = "control", control  = c("PBS"), fdr.type = "Storey's qvalue")
dep <- add_rejections(diff, alpha = 0.01,lfc = 2)

# Get results
results <- get_results(dep)
colnames(results)

significant_proteins <- results[results$significant,]
head(significant_proteins)
}
