% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_result.R
\name{get_signicant}
\alias{get_signicant}
\title{Extract significant candidates from SummarizedExperiment or DEGdata}
\usage{
get_signicant(
  object,
  contrasts = NULL,
  thresholdmethod = NULL,
  curvature = 1,
  x0_fold = 2,
  return_type = c("subset", "table", "names")
)
}
\arguments{
\item{object}{a SummarizedExperiment or DEGdata object.}

\item{contrasts}{NULL or contrasts in object.}

\item{thresholdmethod}{NULL or 'intersect' or 'curve'. The thresholdmethod to decide significant.
If is NULL, used existing rejections. Otherwise filter new rejections via
\code{\link{add_rejections}}().}

\item{curvature}{Numeric(1), Sets the curvature for the curve cutoff lines}

\item{x0_fold}{Numeric(1), decide the x0 ratio to the standard deviations of L2FC. The x0 usually is set to 1(medium confidence) or 2(high confidence) standard deviations.}

\item{return_type}{One of "subset", "table", "names". "subset" return a subset object, "table" return a result data.frame,
names return name vector of object.}
}
\value{
A object in the same class of input or a data.frame or a names vector, designed by return_type.
}
\description{
Extract significant candidates from SummarizedExperiment or DEGdata
}
