% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSimp.R
\name{GS_imp_wrapper}
\alias{GS_imp_wrapper}
\title{pre_processing_GS_wrapper}
\usage{
GS_imp_wrapper(
  data,
  log = F,
  iters_each = 50,
  iters_all = 10,
  lo = -Inf,
  hi = NULL,
  hi_q = 0.15,
  n_cores = 1
)
}
\arguments{
\item{data}{a assay}
}
\description{
import from GSimp github.com/WandeRum/GSimp
}
