% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_enrichment.R
\name{ID_transform}
\alias{ID_transform}
\title{ID transform SE or DEGdata}
\usage{
ID_transform(
  x,
  from_columns = "rownames",
  fromtype = "ENSEMBL",
  species = "Human",
  replace_rowname = "SYMBOL"
)
}
\arguments{
\item{x}{SummarizedExperiment object from \code{\link{make_se}()}) or \code{\link{make_pe}()}),
or a DEGdata object from \code{\link{test_diff_deg}()}).}

\item{from_columns}{Character(), the origin ID from one of "rownames" or column of \code{rowData(x)}}

\item{fromtype}{Character(1), the type of origin ID, e.g. "ENSEMBEL", "SYMBOL", "UNIPROT", "ENTREZID".}

\item{species}{Character(1), the species name.}

\item{replace_rowname}{NULL or character. Should be one of NULL, "SYMBOL", "ENTREZID", "UNIPROT", "ENSEMBL".
If it is not NULL, replace rownames(x) by selected kind of ID.}
}
\value{
A SummarizedExperiment or DEGdata which rowData includes gene information like
"ENSEMBEL", "SYMBOL", "UNIPROT" or "ENTREZID" transformed from giving ID.
If x is a DEGdata, gene information table is also store in the geneinfo slot.
Rownames of object are replaced according replace_rowname.
}
\description{
Transform ID for SummarizedExperiment or DEGdata according origin ID.
The annotation package of certain species must be installed. Using \code{annoSpecies_df}
to check species names and
}
\examples{
\dontrun{
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols,mode = "delim")

## transform uniport ID
check_organismDB_depends(organism = "Mouse") # check annotation package of Mouse
head(rowData(se)$ID)
se2 <- ID_transform(se,from_columns = "ID",fromtype = "UNIPROT",species = "Mouse")
head(rowData(se2)$SYMBOL)
head(rownames(se2))
}
}
