% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_aggregation.R
\name{impute_pe}
\alias{impute_pe}
\title{Impute a QFeatures object}
\usage{
impute_pe(
  pe,
  fun = c("QRILC", "bpca", "knn", "MLE", "MinDet", "MinProb", "man", "min", "zero",
    "mixed", "nbavg", "RF", "GSimp"),
  i = "peptideRaw",
  name = "peptideImp",
  ...
)
}
\arguments{
\item{pe}{A QFeature object}

\item{fun}{Character(1), imputation strategy, see \link{impute}.}

\item{i}{A numeric or a character giving the index or the name, respectively, of the assay(s) to be processed.}

\item{name}{Character(1), naming the new normalized assay. defaule is "peptideImp".}

\item{...}{Other parameters to be passed to \link{impute} function.}
}
\value{
A QFeatures with a imputed assay in giving \code{name}.
}
\description{
Impute the \code{i} assay in QFeatures object though \link{impute} function.
The normalized assay is saved in the \code{name} assay.
}
\examples{
\dontrun{
# Load example peptide data
data(Silicosis_peptide)
ecols <- grep("Intensity.", colnames(Silicosis_peptide), value = TRUE)

# Construct QFeature object
pe_peptides <- make_pe_parse(Silicosis_peptide, columns = ecols, remove_prefix = TRUE, log2transform = TRUE,mode = "delim")

# Filter, imputation
pe <- filter_pe(pe_peptides, thr = 1,fraction = 0.4, filter_formula = ~ Reverse != '+' & Potential.contaminant !="+" )
pe <- impute_pe(pe, fun = "QRILC")
}
}
