% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_rna.R
\name{make_dds}
\alias{make_dds}
\title{Count matrix to DESeqDataSet conversion using an experimental design}
\usage{
make_dds(countData, expdesign, design = as.formula(~condition))
}
\arguments{
\item{countData}{Gene counts matrix or a data.frame contain counts data.}

\item{expdesign}{Data.frame, Experimental design with 'label', 'condition' and 'replicate' information.}

\item{design}{The design formula that transmitted to \code{\link[DESeq2]{DESeqDataSet}}}
}
\value{
A DESeqDataSet object.
}
\description{
make_dds creates a DESeqDataSet object from a gene counts matrix and experimental design table.
The gene identifiers should in the first column of countData.
}
\examples{

\dontrun{
data(Silicosis_count)
colnames(Silicosis_count)

expdesign = get_exdesign_parse(colnames(Silicosis_count))
dds = make_dds(countData = Silicosis_count, expdesign = expdesign)

## DESeq2 way, be same.
# DESeqDataSetFromMatrix(countData = Silicosis_count,colData = expdesign,
                         design = as.formula("~ condition"))
                        )
}


}
