% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_rna.R
\name{make_dds_parse}
\alias{make_dds_parse}
\title{Count matrix to DESeqDataSet conversion by parsing from column names}
\usage{
make_dds_parse(
  countData,
  mode = c("char", "delim"),
  chars = 1,
  sep = "_",
  remove_prefix = F,
  design = as.formula(~condition)
)
}
\arguments{
\item{countData}{Gene counts matrix or a data.frame contain counts data.}

\item{mode}{"char" or "delim",
The mode of parsing the column headers.
"char" will parse the last number of characters as replicate number
and requires the 'chars' parameter.
"delim" will parse on the separator and requires the 'sep' parameter.}

\item{chars}{Numeric(1),
The number of characters to take at the end of the column headers
as replicate number (only for mode = "char").}

\item{sep}{Character(1),
The separator used to parse the column header
(only for mode = "delim").}

\item{remove_prefix}{remove the prefix of sample labels.}

\item{design}{The design formula that transmitted to \code{\link[DESeq2]{DESeqDataSet}}}
}
\value{
A DESeqDataSet object.
}
\description{
make_dds_parse creates a DESeqDataSet object from a gene counts matrix.
The gene identifiers should in the first column of countData.
}
\details{
Column name is splitted by delim or from the tail character(mode = "char").
Detail rule can check \code{\link{get_exdesign_parse}})
}
\examples{
\dontrun{
data(Silicosis_count)
colnames(Silicosis_count)

dds = make_dds_parse(Silicosis_count, mode = "delim", sep = "_")
colData(dds)
}


}
