% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_protein.R
\name{make_se}
\alias{make_se}
\title{Data.frame to SummarizedExperiment object
conversion using an experimental design}
\usage{
make_se(proteins_unique, columns, expdesign, log2transform = TRUE)
}
\arguments{
\item{proteins_unique}{Data.frame,
Protein table with unique names annotated in the 'name' column
(output from \code{\link{make_unique}()}).}

\item{columns}{Integer vector,
Column numbers indicating the columns containing the assay data.}

\item{expdesign}{Data.frame,
Experimental design with 'label', 'condition'
and 'replicate' information.}

\item{log2transform}{Logical(1), whether log2 transform the assay, default TRUE.}
}
\value{
A SummarizedExperiment object
with log2-transformed values.
}
\description{
\code{make_se} creates a SummarizedExperiment object
based on two data.frames: the protein table and experimental design.
}
\examples{
# Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
columns <- grep("LFQ.", colnames(data_unique))

## Load experiement design
data(Silicosis_ExpDesign)
exp_design <- Silicosis_ExpDesign
se <- make_se(data_unique, columns, exp_design)
}
