% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_peptide.R
\name{make_unique_ptm}
\alias{make_unique_ptm}
\title{Make unique names for a modication-enriched peptide table}
\usage{
make_unique_ptm(
  PTMdata,
  gene_name = "Gene.names",
  protein_ID = "Protein",
  aa = "Amino.acid",
  pos = "Position",
  modified_name = NULL,
  delim = ";"
)
}
\arguments{
\item{PTMdata}{Data.frame, modified peptide table
for which unique names and ID for each modified peptide will be created.}

\item{gene_name}{Character(1),
Name of the column containing gene symbol or protein name.}

\item{protein_ID}{Character(1),
Name of the column containing protein ID or gene ID.}

\item{aa}{Character(1),
Name of the column containing abbreviation of modified amino acid, e.g. "K", "S", "T".
It is unnecessary if \code{modified_name} is provided.}

\item{pos}{Character(1),
Name of the column containing position of modification on proteins.
It is unnecessary if \code{modified_name} is provided.}

\item{modified_name}{Character(1),
Name of the column containing names of modified peptide in the format
(gene name)_(modified amino acid)(position of modification), e.g. "TBCA_K51".}

\item{delim}{Character(1),
Sets the delimiter separating the multiple feature names within one protein group(gene_name and protein_ID).}
}
\value{
A data.frame with the 6 additional variables
"name" and "ID" containing unique names and identifiers or modified peptide respectively, e.g. "TBCA_K51" and "O75347_K51".
And "gene_name", protein_ID" save protein information of modified peptide, e.g. "TBCA" and "O75347".
"modified_aa" and "modified_pos" is the modified amino acid and position, e.g. "K" and "51".
}
\description{
\code{make_unique_ptm} generates unique identifiers
for a proteomics dataset based on "modified_name" columns.
If "modified_name" is absent, it will creat a modified name
based on "gene_name"/"Protein", "aa" and "pos".
}
\details{
If modified_name is provide, gene_name, protein_ID, aa and pos are unnessary.
Additional variables is generated by parsing modified_name.
else gene_name or protein_ID, aa and pos are nessary.
}
