% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_SE.R
\name{NAiszero}
\alias{NAiszero}
\title{Transform NA to 0 in assay of a SummarizedExperiment}
\usage{
NAiszero(se)
}
\arguments{
\item{se}{A SummarizedExperiment object which contain NAs in assay.}
}
\value{
A SummarizedExperiment object
}
\description{
Transform NA to 0 in assay of a SummarizedExperiment
}
