% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_aggregation.R
\name{normalize_pe}
\alias{normalize_pe}
\title{Normalize a QFeatures object}
\usage{
normalize_pe(
  pe,
  method = c("diff.median", "quantiles", "quantiles.robust", "vsn"),
  i = "peptideRaw",
  name = "peptideNorm"
)
}
\arguments{
\item{pe}{A QFeature object}

\item{method}{Character(1), normalisation method, one of "diff.median", "quantiles", "quantiles.robust" or "vsn".}

\item{i}{A numeric vector or a character vector giving the index or the name, respectively, of the assay(s) to be processed.}

\item{name}{Character(1) naming the new normalized assay name.}
}
\description{
Normalize a QFeatures object though \link[QFeatures:QFeatures-processing]{QFeatures::normalize} function
}
