% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_protein.R
\name{normalize_vsn}
\alias{normalize_vsn}
\title{Normalization using vsn}
\usage{
normalize_vsn(se)
}
\arguments{
\item{se}{SummarizedExperiment,
Proteomics data (output from \code{\link{make_se}()} or
\code{\link{make_se_parse}()}). It is adviced to first remove
proteins with too many missing values using \code{\link{filter_se}()}.}
}
\value{
A normalized SummarizedExperiment object.
}
\description{
\code{normalize_vsn} performs variance stabilizing transformation
using the \code{\link[vsn]{vsn-package}}.
}
\examples{
# Load example
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
ecols <- grep("LFQ.", colnames(data_unique))
exp_design <- Silicosis_ExpDesign
se <- make_se(data_unique, ecols, exp_design)
se <- make_se_parse(data_unique, ecols, mode = "delim", sep = "_")

# Filter and normalize
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
norm <- normalize_vsn(filt)
}
