% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_rna.R
\name{ntf_deg}
\alias{ntf_deg}
\title{Assign normalized assay for DEGdata object}
\usage{
ntf_deg(DEGdata, f = log2, pc = 1)
}
\arguments{
\item{DEGdata}{a DEGdata object from get_res}

\item{f}{a function to apply to normalized counts}

\item{pc}{a pseudocount to add to normalized counts}
}
\value{
A DEGdata with a normalized assay in ntf slot
}
\description{
\code{ntf_deg} transform a log2(default) transformed normalized assay based on \code{\link[DESeq2]{normTransform}()}
}
