% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_protein.R
\name{Order_cols}
\alias{Order_cols}
\title{Set order of experiment design for SummarizedExperiment}
\usage{
Order_cols(object, order, order_column = "condition")
}
\arguments{
\item{object}{a SummarizedExperiment object}

\item{order}{Character vector, the order to set.}

\item{order_column}{Set order on which column of experiment design, e.g. "condition", "label"}
}
\value{
A object sorted by giving order
}
\description{
\code{Order_cols} can re-set the order of a SummarizedExperiment object according  colData()
}
