% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DEP1.R
\name{plot_cond}
\alias{plot_cond}
\title{Plot frequency of significant conditions per protein
and the overlap in proteins between conditions}
\usage{
plot_cond(dep, plot = TRUE)
}
\arguments{
\item{dep}{SummarizedExperiment,
Data object for which differentially enriched proteins are annotated
(output from \code{\link{test_diff}()} and \code{\link{add_rejections}()}).}

\item{plot}{Logical(1),
If \code{TRUE} (default) the barplot is produced.
Otherwise (if \code{FALSE}), the data which the
barplot is based on are returned.}
}
\value{
A histogram (generated by \code{\link[ggplot2]{ggplot}})
}
\description{
\code{plot_cond} generates a histogram of
the number of proteins per condition and stacks for overlapping conditions.
The function is inherited from DEP package.
}
\examples{
# Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols, mode = "delim", sep = "_")

# Analysis
filt <- filter_se(se, thr = 0, fraction = 0.3, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinDet")
diff <- test_diff(imputed,type = "control", control = "PBS")
dep <- add_rejections(diff)

# Plot histogram with overlaps
plot_cond(dep)
}
