% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_QC.R
\name{plot_cvs}
\alias{plot_cvs}
\title{Plot sample coefficient of variation whitin group}
\usage{
plot_cvs(se)
}
\arguments{
\item{se}{a SummarizedExperiment data object}
}
\value{
a histogram plot of row coefficient of variation (standard deviation / mean) in each condition
}
\description{
Plot sample coefficient of variation whitin group
}
