% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_result.R
\name{plot_diff_hist}
\alias{plot_diff_hist}
\title{Fit a Gaussian distribution for L2FC of each contrast}
\usage{
plot_diff_hist(object, contrasts = NULL, plot = T)
}
\arguments{
\item{object}{a SummarizedExperiment or DEGdata output from \code{test_diff} or \code{test_diff_deg}}

\item{contrasts}{Character, the contrast in object}

\item{plot}{Logical(1), if TRUE return histogram(s) of L2FC in each contras with fitted normal curve.
if FALSE return the fitting result, mu and sigma of Gaussian distribution}
}
\description{
\code{plot_diff_hist} fit a Gaussian distribution for L2FC in each contrast.
The L2FC value is firstly filter out outlier, and then fit to normal curve based on likelihood.
}
