% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_QC.R
\name{plot_imputation}
\alias{plot_imputation}
\title{Visualize imputation}
\usage{
plot_imputation(se, ...)
}
\arguments{
\item{se}{SummarizedExperiment,
Data object, e.g. before imputation
(output from \code{\link{normalize_vsn}()}).}

\item{...}{Other SummarizedExperiment object(s),
E.g. data object after imputation
(output from \code{\link{impute}()}).}
}
\value{
Density plots of all conditions
of all conditions for input objects, e.g. before and
after imputation (generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{plot_imputation} generates density plots
of all conditions for input objects, e.g. before and after imputation.
}
\examples{
# Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Construct SE
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols,mode = "delim")

# Filter and normalization
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
norm <- normalize_vsn(filt)

# imputation
imputed <- impute(norm, fun = "MinProb", q = 0.01)

# Plot imputation
plot_imputation(filt, norm, imputed)
}
