% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_explore.R
\name{plot_ma_RNA}
\alias{plot_ma_RNA}
\title{MA-plot of RNA expression data}
\usage{
plot_ma_RNA(
  x,
  contrast = get_contrast(x)[1],
  point_alpha = 0.2,
  sig_color = "red3",
  hlines = NULL,
  title = contrast,
  xlab = "mean of normalized counts - log10 scale",
  ylim = NULL,
  add_rug = FALSE,
  intgenes = NULL,
  intgenes_color = "steelblue4",
  intgenes_size = 3,
  label_fontsize = 3,
  labels_repel = FALSE
)
}
\arguments{
\item{x}{DEGdata object,
Test result for which differentially expressed/enriched proteins or genes are annotated
(output from \code{\link{test_diff}()}).}

\item{contrast}{Character.
Only plot on certain contrasts.}

\item{point_alpha}{Numeric(1), between 0 to 1,
transparency value for the points (0 = transparent, 1 = opaque)}

\item{sig_color}{Charecter(1),
Color to use to mark differentially expressed genes. Defaults to red}

\item{hlines}{Numeric(1),
The y coordinate (in absolute value) where to draw horizontal lines,
optional}

\item{title}{A title for the plot, optional}

\item{xlab}{X axis label, defaults to "mean of normalized counts - log10 scale"}

\item{ylim}{Vector of two numeric values, Y axis limits to restrict the view.}

\item{add_rug}{Logical, whether to add rug plots in the margins}

\item{intgenes}{Vector of genes of interest, highlighted on plot by \code{intgenes_color}.
Gene symbols if a \code{symbol} column is provided in \code{x},
or else the identifiers specified in the row names}

\item{intgenes_color}{Character(1), The color to use to mark the genes on the main plot.}

\item{intgenes_size}{The point size of \code{intgenes}.}

\item{label_fontsize}{Numeric(1),
The fontsize of \code{labels_intgenes}}

\item{labels_repel}{Logical, whether to use \code{geom_text_repel} for placing the
labels on the features to mark}

\item{labels_intgenes}{Logical, whether to add the gene identifiers/names close
to the marked plots.}
}
\value{
A ggplot object
}
\description{
Plot MA-plot for a differential test result out from \code{\link{test_diff}()}.
}
