% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_explore.R
\name{plot_norm_distribution}
\alias{plot_norm_distribution}
\title{Plot the fit normal for log2 fold change}
\usage{
plot_norm_distribution(object, contrast = get_contrast(object))
}
\arguments{
\item{object}{A summarizedExperiment object}

\item{contrast}{The contrasts to plot fitting normal curve.}
}
\value{
A histogram plot with a normal curve
}
\description{
Plot the fit normal for log2 fold change
}
