% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_result.R
\name{plot_single}
\alias{plot_single}
\title{Plot values for a protein/gene of interest}
\usage{
plot_single(object, proteins, type = c("contrast", "centered"), plot = TRUE)
}
\arguments{
\item{object}{SummarizedExperiment,
Data object for which differentially enriched proteins are annotated
(output from \code{\link{test_diff}()} and \code{\link{add_rejections}()}).}

\item{proteins}{Character,
The name(s) of the protein(s)/gene(s) to plot.}

\item{type}{'contrast' or 'centered',
The type of data scaling used for plotting.
Either the fold change ('contrast') or
the centered log2-intensity ('centered').}

\item{plot}{Logical(1),
If \code{TRUE} (default) the barplot is produced.
Otherwise (if \code{FALSE}), the summaries which the
barplot is based on are returned.}
}
\value{
A barplot (generated by \code{\link[ggplot2]{ggplot}}).
}
\description{
\code{plot_single} generates a barplot of a protein/gene of interest.
}
\examples{
# Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
ecols <- grep("LFQ.", colnames(data_unique))


## Load experiement design
data(Silicosis_ExpDesign)
exp_design <- Silicosis_ExpDesign
se <- make_se(data_unique, ecols, exp_design)

# Filter and normalize
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")

norm <- normalize_vsn(filt)

# Impute missing values using different functions
imputed <- impute(norm, fun = "MinProb", q = 0.05)

# Test for differentially expressed proteins
diff <- test_diff(imputed, "control", "PBS")
dep <- add_rejections(diff, alpha = 0.05, lfc = 1)

# Plot single proteins
plot_single(dep, 'Oas2')
plot_single(dep, 'Oas2', 'centered')
plot_single(dep, c('Oas2', 'Oas3'))
plot_single(dep, c('Oas2', 'Oas3'), plot = FALSE)
}
