% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_result.R
\name{plot_statistics}
\alias{plot_statistics}
\title{Plot the distribution of statistic valuse. develop from statistics_plot}
\usage{
plot_statistics(
  object,
  statistic = c("diff", "t.stastic", "p.val", "p.adj"),
  contrasts = NULL
)
}
\arguments{
\item{object}{A SummarizedExperiment or DEGdata obejct from \code{\link{test_diff}} / \code{\link{test_diff_deg}}
or \code{add_rejections}}

\item{statistic}{Character, in c("diff","t","p.val","p.adj")}

\item{contrasts}{Characters or NULL, plot contrasts. If NULL, plot all tested contrasts in objest.}
}
\value{
The distribution histogram of statistics
}
\description{
Plot the distribution of statistic valuse. develop from statistics_plot
}
