% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_protein.R
\name{reshape_long2wide}
\alias{reshape_long2wide}
\title{Reshape a long table to wide}
\usage{
reshape_long2wide(
  long_table,
  feature_col,
  expression_col,
  sample_col,
  remove_sample_prefix = T,
  remove_sample_suffix = T,
  shrink_ident_cols = NULL,
  extend_ident_cols = NULL
)
}
\arguments{
\item{long_table}{data.frame, a long-format table}

\item{feature_col}{character(1), the unique identifier of feature, such as "protein.group" for protein,
"peptide.sequence" for peptides, "precursor.Id" for precursors.}

\item{expression_col}{character(1), the expression column.}

\item{sample_col}{character(1), the samples column. The samples in these column will be the columns names of output transformed table}

\item{remove_sample_prefix}{logical(1), whether to remove the prefix of samples.}

\item{remove_sample_suffix}{logical(1), whether to remove the suffix of samples.}

\item{shrink_ident_cols}{NULL or characters,the variablesvariables (identification information, like score, protein.names)
to stored in transformed table.
If the variable is multiple for a feature, it will paste to a character string. Else, it will store as the unique value.}

\item{extend_ident_cols}{NULL or characters,the variables (identification information, like score, protein.names)
to stored in transformed table, like the .
The variables will be extended to wide table like the expression_col, the value of different sample will store in different colmun in out table.}
}
\value{
A data.frame in a wide format
}
\description{
Transform the long-format table to wide-format, and each row is feature and cols are expression data or
identification information. The output table can be used as the input of DEP2
}
\examples{
# Read in a example long export table
long_table <- read.csv(system.file("extdata/DIA-NN_Export.tsv.gz",package = "DEP2"),sep = "\t")
head(long_table)

# Reshape to a wide expression table
wide_table <- reshape_long2wide(long_table,sample_col = "File.Name",
                                feature_col = "Precursor.Id", expression_col = "Precursor.Normalised")
head(wide_table)

# If do not remove prefix or suffix of samples
wide_table2 <- reshape_long2wide(long_table,sample_col = "File.Name",
                                 feature_col = "Precursor.Id", expression_col = "Precursor.Normalised",
                                 remove_sample_prefix = FALSE, remove_sample_suffix = FALSE)
colnames(wide_table2)

# Keep some identification information
wide_table2 <- reshape_long2wide(long_table,sample_col = "File.Name",
                                 feature_col = "Precursor.Id", expression_col = "Precursor.Normalised",
                                 shrink_ident_cols = c("Protein.Names","Protein.Group",
                                                       "Stripped.Sequence","Modified.Sequence",
                                                       "Precursor.Charge","Evidence"),
                                 extend_ident_cols = "Q.Value"
)
str(wide_table2)

}
