% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_rna.R
\name{rlg_deg}
\alias{rlg_deg}
\title{Assign a rlog transformed assay for DEGdata object}
\usage{
rlg_deg(DEGdata, blind = FALSE, ...)
}
\arguments{
\item{DEGdata}{a DEGdata object from get_res}

\item{blind}{logical, whether to blind the transformation to the experimental
design. blind=TRUE should be used for comparing samples in an manner unbiased by
prior information on samples, for example to perform sample QA (quality assurance).
blind=FALSE should be used for transforming data for downstream analysis,
where the full use of the design information should be made.
blind=FALSE will skip re-estimation of the dispersion trend, if this has already been calculated.
If many of genes have large differences in counts due to
the experimental design, it is important to set blind=FALSE for downstream
analysis.}

\item{...}{Other options in \code{\link[DESeq2]{rlog}()}.}
}
\value{
A DEGdata with a rlog transform assay in rlg slot
}
\description{
\code{rlg_deg} calulate a rolg transformed assay based on \code{\link[DESeq2]{rlog}()}
}
