% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Silicosis_ExpDesign}
\alias{Silicosis_ExpDesign}
\title{Silicosis_ExpDesign - experiment design of silicosis mouse proteome}
\format{
A data.frame with 20 rows and 4 variables:
\describe{
\item{label/ID}{The label of experiment groups}
\item{condition }{The experiment condition}
\item{replicate}{Replicate number}
}
}
\usage{
Silicosis_ExpDesign
}
\description{
The ExpDesign table of \code{Silicosis_pg} annotating samples information.
}
\keyword{datasets}
