% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Silicosis_peptide}
\alias{Silicosis_peptide}
\title{Silicosis_peptide - peptide quantity data of a silicosis mouse model}
\format{
A data.frame with 143528 rows and 29 variables:
\describe{
\item{Sequence}{Peptide Sequence}
\item{Missed.cleavages}{Missed cleavages number in peptide}
\item{Proteins}{The uniprot ID of proteins that peptide affiliated to}
\item{Leading.razor.protein}{The leading razor protein uniprot ID that peptide affiliated to}
\item{Gene.names}{Gene name of the proteins that peptide affiliated to}
\item{PEP}{Posterior error probability of the identification}
\item{Score}{Andromeda score in Maxquant}
\item{Intensity columns}{Raw intensity of peptides}
\item{Reverse}{The peptide is identified in the decoy database if marked ('+')}
\item{Potential.contaminant}{The peptide from a known contaminant protein if marked ('+')}
}
}
\source{
Wang M, Zhang Z, Liu J, et al.
Gefitinib and fostamatinib target EGFR and SYK to attenuate silicosis:
a multi-omics study with drug exploration. Signal Transduct Target Ther. 2022;7(1):157.
\href{https://www.nature.com/articles/s41392-022-00959-3}{doi: 10.1038/s41392-022-00959-3}.
}
\usage{
Silicosis_peptide
}
\description{
The Silicosis_peptide is peptide quantity of a proteomics study on silica-exposed mouse model
reported by Wang, Zhang, Liu, et al.
}
\keyword{datasets}
