% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Silicosis_pg}
\alias{Silicosis_pg}
\title{Silicosis_pg - proteinGroups data of a silicosis mouse model}
\format{
A data.frame with 10718 rows and 55 variables:
\describe{
\item{Protein.IDs}{Uniprot IDs}
\item{Majority.protein.IDs}{Uniprot IDs of major protein(s) in the protein group}
\item{Peptide.counts..all.}{Total peptides counts of each protein}
\item{Peptide.counts..razor.unique.}{Razor or unique peptides counts of each protein}
\item{Peptide.counts..unique.}{Unique peptides counts of each protein}
\item{Protein.names}{Full protein names}
\item{Gene.names}{Gene name}
\item{Fasta.headers}{Header as present in the Uniprot fasta file}
\item{Peptides}{Number of peptides identified for this protein group}
\item{Razor...unique.peptides}{Number of peptides used for the quantification of this protein group}
\item{Unique.peptides}{Number of peptides identified which are unique for this protein group}
\item{iBAQ columns}{intensity-based absolute-protein-quantification}
\item{LFQ.intensity columns}{LFQ normalized mass spectrometry intensity, A.U.}
\item{Only.identified.by.site}{The protein is only identified by a modification site if marked ('+')}
\item{Reverse}{The protein is identified in the decoy database if marked ('+')}
\item{Potential.contaminant}{The protein is a known contaminant if marked ('+')}
}
}
\source{
Wang M, Zhang Z, Liu J, et al.
Gefitinib and fostamatinib target EGFR and SYK to attenuate silicosis:
a multi-omics study with drug exploration. Signal Transduct Target Ther. 2022;7(1):157.
\href{https://www.nature.com/articles/s41392-022-00959-3}{doi: 10.1038/s41392-022-00959-3}.
}
\usage{
Silicosis_pg
}
\description{
The Silicosis_pg is proteinGroups result of a proteomics study on silica-exposed mouse model
reported by Wang, Zhang, Liu, et al.
}
\keyword{datasets}
