% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Silicosis_phos}
\alias{Silicosis_phos}
\title{Silicosis_phos - phosphoproteome result of a silicosis mouse model}
\format{
A data.frame with 21249 rows and 40 variables:
\describe{
\item{Proteins}{The proteins uniprot IDs that phosphopeptide affiliated to}
\item{Protein}{The first protein uniprot ID}
\item{Gene.names}{Gene names of the proteins that phosphopeptide affiliated to}
\item{Localization.prob}{The localization probability of the phosphorylated site}
\item{PEP}{Posterior error probability of the identification}
\item{Score}{Andromeda score of peptide identification}
\item{Score.for.localization}{Andromeda score of the MS/MS spectrum used for calculating the localization score for this site}
\item{Amino.acid}{The phosphorylated Amino acid}
\item{Intensity columns}{Raw intensity of phosphopeptide}
\item{Reverse}{The peptide is identified in the decoy database if marked ('+')}
\item{Potential.contaminant}{The peptide from a known contaminant protein if marked ('+')}
\item{Positions}{The possible phosphorylation positions in peptides}
\item{Position}{The most possible position with phosphorylation}
}
}
\source{
Wang M, Zhang Z, Liu J, et al.
Gefitinib and fostamatinib target EGFR and SYK to attenuate silicosis:
a multi-omics study with drug exploration. Signal Transduct Target Ther. 2022;7(1):157.
\href{https://www.nature.com/articles/s41392-022-00959-3}{doi: 10.1038/s41392-022-00959-3}.
}
\usage{
Silicosis_phos
}
\description{
The enriched phosphopeptide quantity data on a silica-exposed mouse model
reported by Wang, Zhang, Liu, et al.
}
\keyword{datasets}
