% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_rna.R
\name{test_diff_deg}
\alias{test_diff_deg}
\title{Differential expression test on a DESeqDataSet}
\usage{
test_diff_deg(
  dds,
  type = c("all", "control", "manual"),
  control = NULL,
  test_contrasts = NULL,
  contrast_upon = "condition",
  filter_ihw = FALSE,
  independentFiltering = TRUE,
  lfcshark = FALSE,
  ...
)
}
\arguments{
\item{dds}{A DESeqDataSet object.}

\item{type}{"control", "all" or "manual",
The type of contrasts that will be tested.
This can be all possible pairwise comparisons ("all"),
limited to the comparisons versus the control ("control"), or
manually defined contrasts ("manual").}

\item{control}{Character(1),
The condition to which contrasts are generated if type = "control"
(a control condition would be most appropriate).}

\item{test_contrasts}{Character,
The contrasts that will be tested if type = "manual".
These should be formatted as "SampleA_vs_SampleB" or
c("SampleA_vs_SampleC", "SampleB_vs_SampleC").}

\item{contrast_upon}{Character, the contrast to bulid upon which column in experimentdesign.}

\item{filter_ihw}{logical(1),Whether to use ihw to perform independent filtering and p-value adjustment, default FALSE. See \link[DESeq2]{results}}

\item{independentFiltering}{logical(1), whether independent filtering should be applied automatically, default TRUE. See \link[DESeq2]{results}}

\item{lfcshark}{logical(1),whether to use lfcshark, default FALSE. See \link[DESeq2]{results}}

\item{...}{Parameters transmitted to \link[DESeq2]{DESeq2}}
}
\description{
\code{test_diff_deg} performs a differential expression test based on
\code{DESeq2} package and format the test results from \link[DESeq2]{DESeq}() and \link[DESeq2]{results}().
}
