% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_enrichment.R
\name{test_GSEA}
\alias{test_GSEA}
\title{GSEA data}
\usage{
test_GSEA(
  x,
  type = c("GO", "KEGG", "REACTOME", "MSigDB"),
  species = "Human",
  contrasts = NULL,
  by_contrast = FALSE,
  topn = NULL,
  pAdjustMethod = "BH",
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Enrich biological functions on significant candidate via a over representation analysis.
}
