% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_enrichment.R
\name{test_ORA}
\alias{test_ORA}
\title{ORA for differenatial test result}
\usage{
test_ORA(
  x,
  contrasts = NULL,
  type = c("GO", "KEGG", "REACTOME"),
  species = "Human",
  by_contrast = FALSE,
  pAdjustMethod = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "fdr",
    "none"),
  ...
)
}
\arguments{
\item{x}{A SummarizedExperiment/DEGdata output from add_adjections or
a charachter vector containing candidate identifier(SYMBOL, EntrezID, UniprotID or ENSEMBL).}

\item{contrasts}{Character, analyse results in which contrasts.}

\item{type}{Character, one of "GO","KEGG","REACTOME". The datasets for enrichment analysis.}

\item{species}{The species name.}

\item{by_contrast}{Logical(1). If true, draw enrichment on each contrast, else draw on the total significant candidates.}

\item{pAdjustMethod}{Character, one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{...}{Other parameters in \code{\link[clusterProfiler]{enricher}()} except the cutoff setting}
}
\value{
A enrichResult or compareClusterResult object according by_contrast.
}
\description{
Enrich biological functions on significant candidate via a over representation analysis.
}
\examples{
\dontrun{
# Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Make SummarizedExperiment
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols,mode = "delim")

# Filter and normalize
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
norm <- normalize_vsn(filt)

# Impute missing values using different functions
imputed <- impute(norm, fun = "MinProb", q = 0.05)

# Test for differentially expressed proteins
diff <- test_diff(imputed, type = "control", control  = c("PBS"), fdr.type = "Storey's qvalue")
dep <- add_rejections(diff, alpha = 0.01,lfc = 2)

# GO enrichment
check_organismDB_depends(organism = "Mouse") # check annotation package of Mouse
res_ora <- test_ORA(dep, contrasts = "W4_vs_PBS", species = "Mouse",type = "GO")
enrichplot::dotplot(res_ora)
}
}
