% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_ppi.R
\name{test_PPI}
\alias{test_PPI}
\title{Protein-protein interaction analysis}
\usage{
test_PPI(
  x,
  contrasts = NULL,
  species = "Human",
  STRING.version = "11.5",
  choose_scores = NULL,
  score_cutoff = 400
)
}
\arguments{
\item{x}{A SummarizedExperiment/DEGdata output from \code{\link{add_adjections}} or
a charachter vector containing candidate identifiers(SYMBOL, EntrezID, UniprotID or ENSEMBL).}

\item{contrasts}{Character, specifies the contrasts to get significant candidates.}

\item{species}{The species name.}

\item{STRING.version}{Character(1),the version of STRING data.}

\item{choose_scores}{NULL or "combined_score" or
a subset of c("neighborhood","fusion","cooccurence","coexpression",
"experimental","database","textmining"). If is NULL or "combined_score",
cutoff is on the scores combining above all evidence, otherwise is on the subset.}

\item{score_cutoff}{A numeric lager than 400, required lowest interaction scores.
A score lowwer than 400 means the interaction is unconfident}
}
\value{
A data.frame of PPI information
}
\description{
\code{test_PPI} search protein-protein interaction on significant candidate
based on a local STRING database.
}
\details{
Identifiers in x are firstly transformed to entrezID and then mapped the relative STRING id.
Next the protein-protein interaction is exctracted from STRING.link.detail.\n
The local STRING data (including protein.info, protein.aliases and protein.links.detailed) is necessary for this function.
If it is absent, \code{test_PPI} will download from \link\link{stringdb-static.org/download/} automatically according
\code{species} and \code{STRING.version}.
}
