% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DEP1.R
\name{theme_DEP1}
\alias{theme_DEP1}
\title{DEP ggplot theme 1}
\usage{
theme_DEP1()
}
\value{
ggplot theme
}
\description{
\code{theme_DEP1} is the default ggplot theme inherited  used for plotting
in DEP package with horizontal x-axis labels.
}
