% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_DEP1.R
\name{theme_DEP2}
\alias{theme_DEP2}
\title{DEP ggplot theme 2}
\usage{
theme_DEP2()
}
\value{
ggplot theme
}
\description{
\code{theme_DEP2} is the ggplot theme from DEP package used for plotting
in DEP package with vertical x-axis labels.
}
