% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_aggregation.R
\name{aggregate_pe}
\alias{aggregate_pe}
\title{Summarize peptide quantity to protein quantity}
\usage{
aggregate_pe(
  pe,
  aggrefun = c("RobustSummary", "medianPolish", "totalMean"),
  aggregate_Peptide_Type = c("Unique + Razor", "Unique"),
  fcol,
  peptide_assay_name = "peptideNorm",
  reserve = "Gene.names"
)
}
\arguments{
\item{pe}{A QFeatures object,
contains the normalized peptide assay}

\item{aggrefun}{A function used for quantitative feature aggregation.
It can be a character in "RobustSummary","medianPolish","totalMean" or a function.
Details see \code{\link[QFeatures]{aggregateFeatures}}}

\item{aggregate_Peptide_Type}{Character in "Unique + Razor" or "unique".
Use what kind of peptides to summarise proteins.}

\item{fcol}{Character(1), defining how to summarise the features. Exist in \code{rowData(pe)}.}

\item{peptide_assay_name}{Character(1), the name of aggregation result assay}

\item{reserve}{Character, the column(s) which will reserve after aggregate.}
}
\value{
a QFeatures object with a new protein aggregation assay.
}
\description{
Aggregation peptide to protein vis 'aggrefun'.
}
