% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_aggregation.R
\name{filter_pe}
\alias{filter_pe}
\title{Filter Qfeatures object on missing values or formula}
\usage{
filter_pe(
  pe,
  thr = NULL,
  missnum = NULL,
  fraction = NULL,
  filter_formula = NULL,
  assay_name = "peptideRaw",
  return_keeprows = F
)
}
\arguments{
\item{pe}{Qfeatures,
Proteomics data (output from \code{\link{make_pe}()} or
\code{\link{make_pe_parse}()}).}

\item{thr}{Integer(1),
Sets the threshold for the allowed max number of missing values
in at least one condition. At least in one condition, missing values number < or = 'thr'.}

\item{missnum}{Integer(1),
Sets the threshold for the allowed max number among all samples,
missing values number < or = 'missnum' in each row.}

\item{fraction}{A numeric from 0 to 1, threshold of quantitative fraction of each feature (each row).
Only the features with a valid quantity ratio exceeding \code{fraction} threshold will be kept.}

\item{filter_formula}{Formula or character(1).
A filter expression, written as a formula.}

\item{assay_name}{Character(1),
the name of which assay to perform filter.}

\item{return_keeprows}{Logical(1), default FALSE, return the trimmed object.
If TURE return the rows number that pass through filter.}
}
\value{
A filtered Qfeatures object,
or a vector of keep rows if \code{return_keeprows} is TRUE
}
\description{
\code{filter_pe} filters a Qfeatures object based on missing values in assay
and provides formula.
The dataset is filtered for proteins that have a maximum of
'thr' missing values in at least one condition; or a maximum of 'missnum' in all samples;
or a filter_formula
}
\examples{
# Load example peptide data
data(Silicosis_peptide)
ecols <- grep("Intensity.", colnames(Silicosis_peptide), value = TRUE)

# Construct a QFeatures object, with a 'peptideRaw' assay
pe_peptides <- make_pe_parse(Silicosis_peptide, columns = ecols, remove_prefix = TRUE, log2transform = TRUE,mode = "delim")
filt_pe <- filter_pe(pe_peptides, thr = 1,fraction = 0.4, filter_formula = ~ Reverse != '+' & Potential.contaminant !="+" )

}
