% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_result.R
\name{get_contrast}
\alias{get_contrast}
\title{Get contrast(s) from SummarizedExperiment or DEGdata}
\usage{
get_contrast(object)
}
\arguments{
\item{object}{A SummarizedExperiment or DEGdata object output from \link{test_diff}()}
}
\value{
Character, the existing contrast(s) in object, which is introduced in the test_diff step.
}
\description{
Get contrast(s) from SummarizedExperiment or DEGdata
}
\examples{
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")
# Make SummarizedExperiment
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols, mode = "delim", sep = "_")

filt <- filter_se(se, thr = 0, fraction = 0.3, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinDet")
diff <- test_diff(imputed,type = "control", control = "PBS")
get_contrast(diff)
diff <- test_diff(imputed,type = "manual", test = c("W9_vs_PBS","W6_vs_PBS"))
get_contrast(diff)
}
