% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_rna.R
\name{get_exdesign_parse}
\alias{get_exdesign_parse}
\title{Construct experiment design basic on colnames parse}
\usage{
get_exdesign_parse(label, mode = c("delim", "char"), chars = 1, sep = "_")
}
\arguments{
\item{label}{Characters, label names(column names) of assay}

\item{mode}{one of "delim", "char"}

\item{chars}{Integer(1). If mode is "char", how many chars in the end of label are 'replicate'}

\item{sep}{Character(1). If mode is "sep", how many chars in the end of label are 'replicate'}
}
\value{
A data.frame contain four columns: label,ID,condition,replicate
}
\description{
Construct experiment design basic on colnames parse
}
