% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_enrichment.R
\name{get_ORA_result}
\alias{get_ORA_result}
\title{Extract significant enrichment base on giving threshold}
\usage{
get_ORA_result(
  ORA_enrichment,
  ont = NULL,
  pvalueCutoff = 0.05,
  qvalueCutoff = 0.2,
  simplify = FALSE,
  simplify.cutoff = 0.7,
  simplify.measure = c("Wang", "Resnik", "Lin", "Rel", "Jiang"),
  simplify.semData = NULL,
  return_table = F
)
}
\arguments{
\item{ont}{One of "ALL", "BP", "MF", "CC"}

\item{pvalueCutoff}{Numeric(1), the p.value cutoff on enrichment result}

\item{qvalueCutoff}{Numeric(1), The qvalue cutoff on enrichment tests}

\item{simplify}{Logical(1), if simplify GO terms by}

\item{return_table}{Logical(1), if true return a enrichResult or a result table}

\item{reat}{The output from test_ORA}
}
\value{
A enrichResult/compareClusterResult of significant enrichment,
or a result table of the significant enrichment if return_table is TRUE.
}
\description{
\code{get_ORA_result} perform a
}
