% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_result.R
\name{get_signicant}
\alias{get_signicant}
\title{Extract significant candidates from SummarizedExperiment or DEGdata}
\usage{
get_signicant(
  object,
  contrasts = NULL,
  thresholdmethod = NULL,
  diff = NULL,
  alpha = NULL,
  curvature = NULL,
  x0_fold = NULL,
  return_type = c("subset", "table", "names", "genelist", "idlist")
)
}
\arguments{
\item{object}{a SummarizedExperiment or DEGdata object.}

\item{contrasts}{NULL or contrasts in object.}

\item{thresholdmethod}{NULL or 'intersect' or 'curve'. The thresholdmethod to decide significant.
If is NULL, used existing rejections. Otherwise filter new rejections via
\code{\link{add_rejections}}().}

\item{diff}{Numeric(1) or NULL, The l2fc threshold. Required if thresholdmethod is 'intersect', default is 1.}

\item{alpha}{Numeric(1) or NULL, The p threshold. Required if thresholdmethod is 'intersect', default is 0.05.}

\item{curvature}{Numeric(1), sets the curvature for the curve cutoff lines.
Required if thresholdmethod is 'curve', default is 0.6.}

\item{x0_fold}{Numeric(1), decides the x0 ratio to the standard deviations of L2FC. The x0 usually is set to 1(medium confidence) or 2(high confidence) standard deviations.
Required if thresholdmethod is 'curve', default is 2.}

\item{return_type}{One of "subset", "table", "names","genelist","idlist". "subset" return a subset object, "table" return a result data.frame,
names return name vector of object. "genelist" and "idlist" require a contrasts input and
return the vector of l2fc, the names of vector are the names or the ids. If contrasts is more than one, return the first input contrasts l2fc;
if contrasts is NULL, return the l2fc of the first contrasts of object.}

\item{change_trend}{Character(1), one of "all","up","down". Exctract all, upregulated(L2FC > 0) or downregulated(L2FC < 0) subset.}
}
\value{
A object in the same class of input or a data.frame or a names vector, designed by return_type.
}
\description{
Extract significant candidates from SummarizedExperiment or DEGdata
}
\examples{
# Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Differential test
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols,mode = "delim")
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")
norm <- normalize_vsn(filt)
imputed <- impute(norm, fun = "MinProb", q = 0.05)
diff <- test_diff(imputed, type = "control", control  = c("PBS"), fdr.type = "Storey's qvalue")
dep <- add_rejections(diff, alpha = 0.01,lfc = 2)

# Signicant subset
(sig <- get_signicant(dep))

# Given threshold
(sig <- get_signicant(dep, contrast = "W4_vs_PBS", alpha = 0.001, diff = 2))

# In table format
sig_df <- get_signicant(dep, return_type = "table")
head(sig_df)

}
