% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_aggregation.R
\name{make_pe}
\alias{make_pe}
\title{Data.frame to QFeatures object conversion using an experimental design}
\usage{
make_pe(
  Peptide,
  ecols,
  protein_col,
  expdesign,
  assay_name = "peptideRaw",
  log2transform = T
)
}
\arguments{
\item{Peptide}{Data.frame,
Peptide table with its belonging proteingroup.}

\item{ecols}{A numeric indicating the indices of the columns
to be used as expression values. Can also be a character indicating the names of the columns.}

\item{protein_col}{An optional character(1) or numeric(1) indicating the column of
peptides' proteins/proteingroups information.}

\item{expdesign}{A data.frame.
Experimental design with 'label', 'condition'
and 'replicate' information.}

\item{assay_name}{An character(1) to name assay in the QFeatures object.}

\item{log2transform}{Logical(1), whether log2 transform the assay, default TRUE.}
}
\value{
A QFeatures object with a peptide assay.
}
\description{
\code{make_pe} creates a QFeatures object
based on two data.frames: the peptide table and experimental design.
}
