% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEP_protein.R
\name{Order_cols}
\alias{Order_cols}
\title{Set order of experiment design for SummarizedExperiment}
\usage{
Order_cols(object, order, order_column = "condition")
}
\arguments{
\item{object}{a SummarizedExperiment object}

\item{order}{Character vector, the order to set.}

\item{order_column}{Set order on which column of experiment design, e.g. "condition", "label"}
}
\value{
A object sorted by giving order
}
\description{
\code{Order_cols} can re-set the order of a SummarizedExperiment object according  colData()
}
\examples{
# Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")
# Make SummarizedExperiment
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols, mode = "delim", sep = "_")
colnames(se)
se <- Order_cols(se,c("PBS","W2","W4","W6","W9","W10"))
colnames(se)

}
