% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_QC.R
\name{plot_missval}
\alias{plot_missval}
\title{Plot a heatmap of proteins with missing values}
\usage{
plot_missval(se, ...)
}
\arguments{
\item{se}{SummarizedExperiment}

\item{...}{Additional arguments for Heatmap function as depicted in
\code{\link[ComplexHeatmap]{Heatmap}}
Data object with missing values.}
}
\value{
A heatmap indicating whether values are missing (0) or not (1)
(generated by \code{\link[ComplexHeatmap]{Heatmap}}).
}
\description{
\code{plot_missval} generates a heatmap of proteins
with missing values to discover whether values are missing by random or not.
}
\examples{
# Load example
data(Silicosis_pg)
data <- Silicosis_pg
data_unique <- make_unique(data, "Gene.names", "Protein.IDs", delim = ";")

# Differential test
ecols <- grep("LFQ.", colnames(data_unique))
se <- make_se_parse(data_unique, ecols,mode = "delim")
filt <- filter_se(se, thr = 0, fraction = 0.4, filter_formula = ~ Reverse != "+" & Potential.contaminant!="+")

# Plot missing values heatmap
plot_missval(filt)

}
