% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function_result.R
\name{plot_multi_heatmap}
\alias{plot_multi_heatmap}
\title{Plot heatmap of specified genes/proteins across multiple omics results}
\usage{
plot_multi_heatmap(
  omics_list,
  choose_name,
  to_upper = FALSE,
  color = c("RdBu", "RdYlBu", "RdYlGn", "BrBG", "PiYG", "PRGn", "PuOr", "RdGy",
    "Spectral"),
  col_limit = 6,
  width = 7,
  height = 10,
  row_font_size = 5,
  col_font_size = 11,
  ...
)
}
\arguments{
\item{omics_list}{A list composed of SummarizedExperiment or DEGdata objects.}

\item{choose_name}{Character, the vector of specified identifiers.}

\item{to_upper}{Logical, whether transform all identifiers to upper}

\item{color}{Character(1), sets the color panel (from \pkg{RColorBrewer}).
It can be a palette name in "RdBu", "RdYlBu", "RdYlGn", "BrBG", "PiYG", "PRGn", "PuOr", "RdGy", "Spectral".}

\item{col_limit}{Integer(1),
Sets the outer limits of the color scale.}

\item{width}{Numeric, the width of heatmap}

\item{height}{Numeric, the height of heatmap}

\item{row_font_size}{Numeric, the size of row label.}

\item{...}{other parameters to \code{\link[ComplexHeatmap]{Heatmap}}}
}
\value{
A HeatmapList from \code{ComplexHeatmap} package
}
\description{
Plot heatmap of specified genes/proteins across multiple omics results
}
