% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_enrichment.R
\name{test_ORA}
\alias{test_ORA}
\title{ORA for differenatial test result}
\usage{
test_ORA(
  x,
  type = c("GO", "KEGG", "REACTOME"),
  species = "Human",
  by_contrast = FALSE,
  pAdjustMethod = c("BH", "holm", "hochberg", "hommel", "bonferroni", "BY", "fdr",
    "none"),
  ...
)
}
\arguments{
\item{x}{A SummarizedExperiment/DEGdata output from add_adjections or
a charachter vector containing candidate identifier(SYMBOL, EntrezID, UniprotID or ENSEMBL).}

\item{type}{Character, one of "GO","KEGG","REACTOME". The datasets for enrichment analysis.}

\item{species}{The species name.}

\item{by_contrast}{Logical(1). If true, draw enrichment on each contrast, else draw on the total significant candidates.}

\item{pAdjustMethod}{Character, one of "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none".}

\item{...}{Other parameters in \code{\link[clusterProfiler]{enricher}()} except the cutoff setting}
}
\value{
A enrichResult or compareClusterResult object according by_contrast.
}
\description{
Enrich biological functions on significant candidate via a over representation analysis.
}
